firebase: Make firebase a module
This give us a chance if we suddenly want to disable Firebase in our build.
